+++ grit-docs/docs/going-further/CallToAction.tsx
@@ -1,9 +1,0 @@
-import Link from "@docusaurus/Link";
-
-export function CallToAction() {
-  return (
-      <Link className="button button--primary button--lg" to="mailto:contact@grit42.com">
-        Get in touch 🚀
-      </Link>
-  );
-}