+++ modules/compounds/backend/app/models/grit/compounds/batch_property.rb
@@ -23,6 +23,7 @@ module Grit::Compounds
     belongs_to :compound_type, optional: true
     belongs_to :data_type, class_name: "Grit::Core::DataType"
     belongs_to :unit, class_name: "Grit::Core::Unit", optional: true
+    has_many :batch_property_values
 
     display_column "name"