+++ packages/frontend/form/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.6.0
+
+### 🚀 Features
+
+- controlled vocabularies as data types
+
+### 🧱 Updated Dependencies
+
+- Updated @grit42/client-library to 0.6.0
+
 ## 0.5.0
 
 ### 🧱 Updated Dependencies