+++ packages/frontend/notifications/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.6.0
+
+### 🚀 Features
+
+- controlled vocabularies as data types
+
 ## 0.5.0
 
 This was a version bump only for @grit42/notifications to align it with other projects, there were no code changes.